refactor(tile_stitch): move parse_timepoints into cli, drop _zarr_util
_zarr_util held only a CLI arg parser after the create_zeros inlining;
fold it into its sole caller and remove the module.
